production_requirements.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.9
django-crispy-forms==1.5.2
django-jet==0.1.3
djangorestframework==3.3.2
Pillow==3.1.0
requests==2.9.1
static3==0.6.1
wheel==0.24.0
whitenoise==2.0.6
